refactor: remove Windows Autopilot Device CSV Import functionality
- Deleted the Windows Autopilot Device CSV Import package, including related files for CSV reading, resource management, and model definitions.
- Updated import paths in `ephemeral_resources.go` to reflect the removal of the package.
- This change streamlines the codebase by eliminating unused components, enhancing maintainability.
